requirements.txt
kafka_python==2.0.2
pandas==2.0.3
pyspark==3.5.1
Requests==2.31.0
